Moved and renamed FlagsView & PrereqsView
