+++ src/renderer/dialogs/ConfigurationDialog.tsx
@@ -33,7 +33,7 @@ import { StepIconProps } from '@mui/material/StepIcon'
 import { ColorlibConnector, ColorlibStepIconRoot } from '../components/Colorlib'
 import AuthView from '../components/Config/AuthView'
 import ConfigsView from '../components/Config/ConfigsView'
-import ConfigFlagsView from '../components/ConfigFlagsView'
+import FlagsView from '../components/Config/FlagsView'
 import ConfigSummaryView from '../components/ConfigSummaryView'
 import ConfigVarsView from '../components/ConfigVarsView'
 
@@ -185,7 +185,7 @@ const ConfigurationDialog = ({ onClose }: Props) => {
       content: (
         <Box sx={{ marginLeft: 2, marginRight: 2 }}>
           <ConfigsView localConfigs={localConfigs} onChange={onChangeConfig} />
-          <ConfigFlagsView localFlags={localFlags} onChange={onChangeFlag} />
+          <FlagsView localFlags={localFlags} onChange={onChangeFlag} />
         </Box>
       )
     },
